requirements.txt
PyPDF2==3.0.1
google-generativeai==0.3.2
openpyxl==3.1.2 
dotenv